pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "daily-brief-bot"
version = "0.1.0"
description = "Daily briefing bot that sends curated news and topics to Slack"
requires-python = ">=3.11"
dependencies = [
    "slack-sdk>=3.27.0",
    "openai>=1.30.0",
    "apscheduler>=3.10.0",
    "feedparser>=6.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
daily-brief = "src.main:main"
